requirements.txt
asgiref==3.8.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.2
cloudinary==1.44.0
cryptography==46.0.5
Django==5.2
django-allauth==65.14.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.2
django-jazzmin==3.0.1
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
geographiclib==2.1
geopy==2.4.1
google-auth==2.48.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pillow==11.2.1
playwright==1.52.0
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyee==13.0.0
PyJWT==2.11.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
selenium==4.41.0
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
